+++ src/plskern.jl
@@ -37,7 +37,7 @@ techniques from nonparametric statistics for the real time
 robot learning. Applied Intell., 17, 49-60.
 
 Sicard, E. Sabatier, R., 2006. Theoretical framework for local 
-PLS1 regression and application to a rainfall data set. Comput. Stat. 
+PLS1 regression and application to a rainfall dataset. Comput. Stat. 
 Data Anal., 51, 1393-1410.
 
 ## Examples